OK-746: Siivottu ja pilkottu koodia pienempiin paloihin
