+++ src/app/(root)/@controls/page.tsx
@@ -7,17 +7,17 @@ import { Tila, getHakuAlkamisKaudet } from '@/app/lib/types/kouta-types';
 import { useHakuSearchParams } from '@/app/hooks/useHakuSearch';
 import { useHakutavat } from '@/app/hooks/useHakutavat';
 import { useTranslations } from '@/app/hooks/useTranslations';
-import { OphFormControl } from '@/app/components/form/oph-form-control';
 import { SpinnerIcon } from '@/app/components/spinner-icon';
 import { LocalizedSelect } from '@/app/components/localized-select';
 import { SearchInput } from '@/app/components/search-input';
+import { OphFormControl } from '@/app/components/form/oph-form-control';
 
 const HakutapaSelect = ({
   labelId,
   value: selectedHakutapa,
   onChange,
 }: {
-  labelId: string;
+  labelId?: string;
   value: string;
   onChange: (e: SelectChangeEvent) => void;
 }) => {
