+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulokset/components/laskennaton-valintatapajono-table.tsx
