+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulokset/lib/refetchLaskennanTulokset.tsx
@@ -1,0 +1,14 @@
+import { hakukohteenLasketutValinnanvaiheetQueryOptions } from '@/app/hooks/useLasketutValinnanVaiheet';
+import { QueryClient } from '@tanstack/react-query';
+
+export const refetchLaskennanTulokset = ({
+  queryClient,
+  hakukohdeOid,
+}: {
+  queryClient: QueryClient;
+  hakukohdeOid: string;
+}) => {
+  const options = hakukohteenLasketutValinnanvaiheetQueryOptions(hakukohdeOid);
+  queryClient.resetQueries(options);
+  queryClient.invalidateQueries(options);
+};
